Default the trip start time to the current time, instead of dying.
Railways' request.get method doesn't take a default value, and returns
None if the requested parameter isn't there.  The date parsing routines
don't care for this much, and threw exceptions if they were passed None.
If the user didn't specify the time, leave it as a blank string, which
gets parsed as the current time.