Show the current time as the default start time.
It's not 100% obvious that you can leave the time field blank to get the
current time.  Instead of trying to document it, just show the current
time there when you load the page.  Select the text when the field gets
focus; this lets you change it easily, but also lets you tab past it to
get to the Plan button without making the text disappear distractingly.