Reword the directions.
We now emphasize the times of various events, as that was getting lost
in the text before.  Also include a separate step for getting off a
bus, to make the individual bus trips more obvious, and to help clarify
any walking to be done at the end, and list the destination explicitly
in the directions.