Serve all web pages via Django, rather than Railways.
Dealing with two web frameworks is confusing, and Django's just as capable
of serving requests as railways is.  Move all the code out of routez.py
and into Django views.  Note that this is still somewhat broken, but at
least the Django server starts.

Note that if you want Django to work, you can't have a file called
routez.py hanging around, it confuses things terribly.