Implement TripAction's copy constructor and assignment operator.
In the absence of an explicit copy constructor and assignment operator,
the compiler was generating one.  This is fine for the basic boring
data members, but taking a bitwise copy of a shared_ptr could get dicey.
We might as well be clear about what's going on.

Also switch the regular constructor to use an initializer list, and add
an empty destructor, just for completeness.