Delay a trip's start time until just before the first bus arrives.
If the best path for a trip involves a delay from the provided start
time to the first bus's arrival time, we used to spend that time waiting
at the bus stop.  This isn't very productive.  Now if there's a delay,
postpone the start of the trip so we arrive at the bus stop a few minutes
before the bus.